Create MainApplication
