+++ app/src/main/java/com/theuhooi/uhooipicbook/MainActivity.kt
@@ -1,8 +1,5 @@
 package com.theuhooi.uhooipicbook
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -26,7 +23,6 @@ class MainActivity : AppCompatActivity(), MonsterListFragment.OnListFragmentInte
 
         setContentView(R.layout.activity_main)
         configureToolBar()
-        createNotificationChannel()
     }
 
     override fun onSupportNavigateUp() = findNavController(R.id.nav_host_fragment).navigateUp()
@@ -59,22 +55,6 @@ class MainActivity : AppCompatActivity(), MonsterListFragment.OnListFragmentInte
         }
         val appBarConfiguration = AppBarConfiguration(navController.graph)
         setupActionBarWithNavController(navController, appBarConfiguration)
-    }
-
-    private fun createNotificationChannel() {
-        createInfoNotificationChannel()
-    }
-
-    private fun createInfoNotificationChannel() {
-        val id = getString(R.string.info_channel_id)
-        val name = getString(R.string.info_channel_name)
-        val importance = NotificationManager.IMPORTANCE_DEFAULT
-        val channel = NotificationChannel(id, name, importance).apply {
-            description = getString(R.string.info_channel_description)
-        }
-        val notificationManager =
-            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(channel)
     }
 
     // endregion
